requirements.txt
requests
schedule
python-dotenv
sqlite3
